Refactor appointment data model and improve dashboard view models
-   Rename `Appoinment.kt` to `Appointment.kt`
-   Delete `Appointment.kt` from `data/model`
-   Refactor `DoctorHomeViewModel` to use the new `Appointment` model and renamed `appointmentsList`
-   Rename `DashboardViewModel` to `StandardDashboard`
-   Update `PatientDashboardViewModel` to use the new `Appointment` model and `appointmentsList`
