+++ app/src/main/java/com/example/e_clinic_app/backend/home/DoctorHomeViewModel.kt
@@ -1,10 +1,9 @@
 package com.example.e_clinic_app.backend.home
 
-import com.example.e_clinic_app.data.model.Appointment
+import com.example.e_clinic_app.data.appointment.Appointment
 import com.google.firebase.firestore.FirebaseFirestore
 
-class DoctorHomeViewModel : StandrdDashboard() {
-
+class DoctorHomeViewModel : StandardDashboard(){
     override fun fetchAppointments(firestore: FirebaseFirestore) {
         firestore.collection("appointments")
             .whereEqualTo("doctorId", userId)
@@ -12,14 +11,17 @@ class DoctorHomeViewModel : StandrdDashboard() {
             .get()
             .addOnSuccessListener { documents ->
                 val appointments = documents.toObjects(Appointment::class.java)
-                listOfAppointments.clear()
-                listOfAppointments.addAll(appointments)
+                super.appointmentsList.clear()
+                super.appointmentsList.addAll(appointments)
                 // Notify observers or update UI as needed
             }
             .addOnFailureListener { exception ->
                 // Log or handle the error appropriately
             }
+
     }
+
+
