+++ app/src/main/java/com/example/e_clinic_app/backend/home/PatientDashboardViewModel.kt
@@ -1,6 +1,6 @@
 package com.example.e_clinic_app.backend.home
 
-import com.example.e_clinic_app.data.appointment.Appoinment
+import com.example.e_clinic_app.data.appointment.Appointment
 import com.google.firebase.firestore.FirebaseFirestore
 
 class PatientDashboardViewModel :StandardDashboard() {
@@ -11,10 +11,10 @@ class PatientDashboardViewModel :StandardDashboard() {
             .whereIn("status", listOf("CONFIRMED", "PENDING"))
             .get()
             .addOnSuccessListener { documents ->
-                val appointments = documents.toObjects(Appoinment::class.java)
-                super.appointments.clear()
-                super.appointments.addAll(appointments)
-                // Notify observers or update UI as needed
+                val appointments = documents.toObjects(Appointment::class.java)
+                super.appointmentsList.clear()
+                super.appointmentsList.addAll(appointments)
+                // Notify observers or update UI as needed0
             }
             .addOnFailureListener { exception ->
                 // Log or handle the error appropriately
