+++ app/src/main/java/com/example/e_clinic_app/backend/home/StandardDashboard.kt
@@ -1,7 +1,7 @@
 package com.example.e_clinic_app.backend.home
 
 import androidx.lifecycle.ViewModel
-import com.example.e_clinic_app.data.model.Appointment
+import com.example.e_clinic_app.data.appointment.Appointment
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FirebaseFirestore
 
@@ -11,7 +11,7 @@ abstract class StandardDashboard : ViewModel() {
     val user = FirebaseAuth.getInstance().currentUser
     val userId = user?.uid
 
-    val appointments = mutableListOf<Appointment>()
+    val appointmentsList = mutableListOf<Appointment>()
 
     abstract fun fetchAppointments(firestore: FirebaseFirestore)
