+++ app/src/main/java/com/example/e_clinic_app/data/appointment/Appointment.kt
@@ -5,7 +5,7 @@ import com.example.e_clinic_app.data.users.patient.Patient
 import com.google.firebase.Timestamp
 import java.util.UUID
 
-data class Appoinment(
+data class Appointment(
     val id: UUID,
     val doctor: Doctor,
     val patient: Patient,
